sample/requirement.txt
aiohttp==3.9.5
beautifulsoup4==4.12.3
newspaper3k==0.2.8
nltk==3.8.1
pandas==2.2.2
pyttsx3==2.90
requests==2.32.3
streamlit==1.35.0
transformers==4.41.2
youtube_transcript_api==0.6.2
googlesearch-python
lxml==4.9.3